# api-sys_produtos/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Produtos-api Documentation",
        "description": "api-sys_produtos",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8989",
            "description": "local server"
        }
    ],
    "paths": {
...
        "/api/user/create": {
            "get": {
                "tags": [
                    "Usuário"
                ],
                "summary": "Create Usuário",
                "description": "Usuário Controller",
                "operationId": "e3c2ca93124f94bc0ea43c2ee3c635e3",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/user/update": {
            "get": {
                "tags": [
                    "Usuário"
                ],
                "summary": "Update Usuário",
                "description": "Usuário Controller",
                "operationId": "d27e23329850982df02843c793d3cc2b",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}